Add TCK abstract tests for GeneratorEntsoeCategory and BusbarSectionPosition
